[CMS-28] refactor
