+++ app/Services/PublishPagesService.php
@@ -7,9 +7,9 @@ class PublishPagesService
 {
     protected $pageRepo;
 
-    public function __construct(PageRespository $pageRespository)
+    public function __construct()
     {
-        $this->pageRepo = $pageRespository;
+        $this->pageRepo = new PageRespository;
     }
 
     public function load($pageId) {
