+++ routes/web.php
@@ -4,7 +4,6 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\PageController;
 use App\Http\Controllers\CustomPagesController;
-use App\Http\Controllers\PublishPagesController;
 
 /*
 |--------------------------------------------------------------------------
@@ -30,4 +29,4 @@ Route::group(['middleware' => ['auth']], function () {
 });
 
 Route::post('/save-page', [CustomPagesController::class, 'store']);
-Route::get('/published/{id}',  [PublishPagesController::class, 'loadPages']);
+Route::get('/published/{id}',  [CustomPagesController::class, 'loadPages']);
